requirements.txt
Flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
Werkzeug==3.0.1
Flask-Login==0.6.3
dnspython==2.5.0
email-validator==2.1.0
